requirements.txt
Adyen==4.0.0
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.0
boto3==1.16.43
botocore==1.19.43
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
Django==3.0.11
django-admin-honeypot==1.1.0
django-colorfield==0.3.2
django-colorful==1.3
django-compat==1.0.15
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-debug-toolbar-request-history==0.1.3
django-dotenv==1.4.2
django-extra-views==0.13.0
django-filter==2.4.0
django-haystack==3.0
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-phonenumber-field==3.0.1
django-storages==1.11.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
drf-yasg==1.20.0
factory-boy==2.12.0
Faker==4.14.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
install==1.3.4
iptools==0.6.1
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
locust==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
ndg-httpsclient==0.5.1
packaging==20.4
phonenumbers==8.12.13
Pillow==8.0.1
psutil==5.8.0
psycopg2-binary==2.8.5
purl==1.5
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyotp==2.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.22
python-monkey-business==1.0.0
pytz==2020.4
pyzmq==22.0.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
